requirements.txt
# requirements.txt

allure-pytest==2.13.5
allure-python-commons==2.13.5
Appium-Python-Client==4.4.0
attrs==24.3.0
certifi==2024.12.14
h11==0.14.0
idna==3.10
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==24.2
pluggy==1.5.0
PySocks==1.7.1
pytest==8.3.4
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
